requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.3
django-admin-interface==0.30.1
django-ckeditor-5==0.2.18
django-colorfield==0.14.0
django-js-asset==3.1.2
fontawesomefree==6.6.0
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
python-slugify==8.0.4
redis==6.2.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
